Basic implementation of task assignment for technicians and payment features.
- Updated assignment flow and job confirmation logic.
- Integrated payment features: transaction creation, sandbox testing, QR code processing.
- Fixed issues related to authentication, payment data encryption, and booking/invoice status updates.
- Refactored several services for better maintainability.

+++ demo/src/main/java/com/example/demo/model/dto/MaintenanceCatalogModelPartRequest.java
@@ -3,6 +3,8 @@ package com.example.demo.model.dto;
 import lombok.*;
 import jakarta.validation.constraints.*;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -13,7 +15,7 @@ public class MaintenanceCatalogModelPartRequest {
     private Long partId;
 
     @Positive(message = "Số lượng phải lớn hơn 0")
-    private Integer quantityRequired;
+    private BigDecimal quantityRequired;
 
     private Boolean isOptional = false;
