+++ demo/src/main/java/com/example/demo/repo/PartRepo.java
@@ -5,6 +5,7 @@ import com.example.demo.model.modelEnum.EntityStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -19,7 +20,7 @@ public interface PartRepo extends JpaRepository<Part, Long> {
 
     List<Part> findByStatus(EntityStatus status);
 
-    List<Part> findByQuantityLessThanEqual(Integer threshold);
+    List<Part> findByQuantityLessThanEqual(BigDecimal threshold);
 
     List<Part> findByNameContainingIgnoreCase(String keyword);
 }
